Rename gc package for initialization ordering
