+++ init_tinygo.go
@@ -5,7 +5,7 @@
 
 package main
 
-import _ "github.com/corazawaf/coraza-proxy-wasm/internal/gc"
+import _ "github.com/corazawaf/coraza-proxy-wasm/internal/agc"
 
 // #cgo LDFLAGS: lib/libinjection.a lib/libre2.a lib/libcre2.a lib/libc++.a lib/libc++abi.a lib/libclang_rt.builtins-wasm32.a lib/libaho_corasick.a lib/libmimalloc.a
 import "C"
