+++ internal/agc/gc_globals.go
@@ -6,7 +6,7 @@
 // Copied from https://github.com/tinygo-org/tinygo/blob/3dbc4d52105f4209ece1332f0272f293745ac0bf/src/runtime/gc_globals.go
 // with private start symbols redefined.
 
-package gc
+package agc
 
 import "unsafe"
